preparing for curseforge-friendly release

+++ Kerbalua/Kerbalua/Gui/ScriptNameInputArea.cs
@@ -55,7 +55,7 @@ namespace Kerbalua.Gui {
 		public void Save(string text)
 		{
 			try {
-				File.WriteAllText(CreateFullPath(), text.Replace("\n",Environment.NewLine));
+				File.WriteAllText(CreateFullPath(true), text.Replace("\n",Environment.NewLine));
 				CommonSaveLoadActions();
 			}
 			catch(Exception e) {
@@ -68,7 +68,8 @@ namespace Kerbalua.Gui {
 			string result = "";
 			try { 
 				CommonSaveLoadActions();
-				result=File.ReadAllText(CreateFullPath()).Replace("\r","");
+				CreateFullPath();
+				result=RedOnion.KSP.ROS.KspRosEngine.LoadScript(content.text).Replace("\r","");
 			}
 			catch(Exception e) {
 				UnityEngine.Debug.Log(e.StackTrace);
@@ -76,7 +77,7 @@ namespace Kerbalua.Gui {
 			return result;
 		}
 
-		string CreateFullPath()
+		string CreateFullPath(bool forSave = false)
 		{
 			if (content.text == "") {
 				content.text = defaultScriptFilename;
@@ -85,7 +86,7 @@ namespace Kerbalua.Gui {
 			Directory.CreateDirectory(Settings.BaseScriptsPath);
 			string fullPath = Path.Combine(Settings.BaseScriptsPath, content.text);
 
-			if (!File.Exists(fullPath)) {
+			if (forSave && !File.Exists(fullPath)) {
 				File.WriteAllText(fullPath, "");
 			}
 
@@ -119,7 +120,7 @@ namespace Kerbalua.Gui {
 		List<string> GetScriptList()
 		{
 			if (scriptList == null || ioDelayWatch.ElapsedMilliseconds > ioDelayMillis) {
-				scriptList = new List<string>(Directory.GetFiles(Settings.BaseScriptsPath));
+				scriptList = RedOnion.KSP.ROS.KspRosEngine.EnumerateScripts();
 				ioDelayWatch.Reset();
 				ioDelayWatch.Start();
 			}
